% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Adding two things}
\usage{
add_two_numbers(x, y)
}
\arguments{
\item{x}{A numeric vector of length 1.}

\item{y}{The same as x. \code{haao}.}
}
\description{
I like adding.
}
\examples{
add_two_numbers(1, 2)
# is the same as
1 + 2
}
